Split out base GraphQL module
