+++ modules/i18n/client-react/i18nModule.tsx
@@ -43,7 +43,7 @@ class RootComponent extends React.Component<Props> {
 
 export default (settings.i18n.enabled
   ? new ClientModule(commonI18n, langPicker, {
-      data: { i18n: true },
+      context: { i18n: true },
       // eslint-disable-next-line react/display-name
       rootComponentFactory: [req => <RootComponent req={req} />]
     })
