+++ modules/module/client-vue/ClientModule.ts
@@ -1,7 +1,7 @@
 import BaseModule, { BaseModuleShape } from './BaseModule';
 
 export interface ClientModuleShape extends BaseModuleShape {
-  module?: any;
+  // TODO: Add proper type
   route?: [];
 }
 
@@ -10,10 +10,6 @@ interface ClientModule extends ClientModuleShape {}
 class ClientModule extends BaseModule {
   constructor(...modules: ClientModuleShape[]) {
     super(...modules);
-  }
-
-  get modules(): any {
-    return this.module.map((module: any) => module);
   }
 
   get routes(): any {
