+++ modules/module/common/index.ts
@@ -1,2 +1,4 @@
 export { default } from './CommonModule';
 export * from './CommonModule';
+export { default as GraphQLModule } from './GraphQLModule';
+export * from './GraphQLModule';
