+++ modules/upload/server-ts/index.ts
@@ -15,7 +15,7 @@ export * from './FileSystemStorage';
 
 export default new ServerModule({
   schema: [schema],
-  data: {
+  context: {
     fileSystemStorage
   },
   createResolversFunc: [createResolvers],
